05.31

+++ src/main/java/com/prj2spring/controller/comment/CommentController.java
@@ -3,6 +3,7 @@ package com.prj2spring.controller.comment;
 import com.prj2spring.domain.comment.Comment;
 import com.prj2spring.service.comment.CommentService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
@@ -20,6 +21,7 @@ public class CommentController {
     @PreAuthorize("isAuthenticated()")
     public ResponseEntity addComment(@RequestBody Comment comment,
                                      Authentication authentication) {
+
         if (service.validate(comment)) {
             service.add(comment, authentication);
 
@@ -27,16 +29,39 @@ public class CommentController {
         } else {
             return ResponseEntity.badRequest().build();
         }
+
     }
 
     @GetMapping("list/{boardId}")
-    public List<Comment> lsit(@PathVariable Integer boardId) {
+    public List<Comment> list(@PathVariable Integer boardId) {
 
         return service.list(boardId);
     }
 
     @DeleteMapping("remove")
-    public void remove(@RequestBody Comment comment) {
-        service.remove(comment);
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity remove(@RequestBody Comment comment,
+                                 Authentication authentication) {
+        if (service.hasAccess(comment, authentication)) {
+            service.remove(comment);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+
+    }
+
+
+    @PutMapping("edit")
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity edit(@RequestBody Comment comment,
+                               Authentication authentication) {
+        if (service.hasAccess(comment, authentication)) {
+            service.update(comment);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 }
